requirements.txt
langgraph
pandas
numpy
google-auth
google-auth-oauthlib
google-api-python-client
google.auth-httplib2
langchain
langchain-ollama
python-dotenv